Move and refactor host team transfer endpoint
- Move API endpoint to `/hosts/transfer`.
- Refactor service and datastore methods from teams to hosts.
