+++ server/datastore/datastore_hosts.go
@@ -14,6 +14,7 @@ import (
 	"github.com/fleetdm/fleet/server/test"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"gopkg.in/guregu/null.v3"
 )
 
 var enrollTests = []struct {
@@ -798,4 +799,51 @@ func testHostByIdentifier(t *testing.T, ds kolide.Datastore) {
 
 	h, err = ds.HostByIdentifier("foobar")
 	require.Error(t, err)
+}
+
+func testAddHostsToTeam(t *testing.T, ds kolide.Datastore) {
+	team1, err := ds.NewTeam(&kolide.Team{Name: "team1"})
+	require.NoError(t, err)
+	team2, err := ds.NewTeam(&kolide.Team{Name: "team2"})
+	require.NoError(t, err)
+
+	for i := 0; i < 10; i++ {
+		test.NewHost(t, ds, fmt.Sprint(i), "", "key"+fmt.Sprint(i), "uuid"+fmt.Sprint(i), time.Now())
+	}
+
+	for i := 1; i <= 10; i++ {
+		host, err := ds.Host(uint(i))
+		require.NoError(t, err)
+		assert.Equal(t, null.Int{}, host.TeamID)
+	}
+
+	require.NoError(t, ds.AddHostsToTeam(&team1.ID, []uint{1, 2, 3}))
+	require.NoError(t, ds.AddHostsToTeam(&team2.ID, []uint{3, 4, 5}))
+
+	for i := 1; i <= 10; i++ {
+		host, err := ds.Host(uint(i))
+		require.NoError(t, err)
+		expectedID := null.Int{}
+		switch {
+		case i <= 2:
+			expectedID = null.IntFrom(int64(team1.ID))
+		case i <= 5:
+			expectedID = null.IntFrom(int64(team2.ID))
+		}
+		assert.Equal(t, expectedID, host.TeamID)
+	}
+
+	require.NoError(t, ds.AddHostsToTeam(nil, []uint{1, 2, 3, 4}))
+	require.NoError(t, ds.AddHostsToTeam(&team1.ID, []uint{5, 6, 7, 8, 9, 10}))
+
+	for i := 1; i <= 10; i++ {
+		host, err := ds.Host(uint(i))
+		require.NoError(t, err)
+		expectedID := null.Int{}
+		switch {
+		case i >= 5:
+			expectedID = null.IntFrom(int64(team1.ID))
+		}
+		assert.Equal(t, expectedID, host.TeamID)
+	}
 }
