+++ server/service/transport_hosts.go
@@ -2,6 +2,7 @@ package service
 
 import (
 	"context"
+	"encoding/json"
 	"net/http"
 )
 
@@ -36,4 +37,13 @@ func decodeListHostsRequest(ctx context.Context, r *http.Request) (interface{}, 
 	}
 
 	return listHostsRequest{ListOptions: hopt}, nil
+}
+
+func decodeAddHostsToTeamRequest(ctx context.Context, r *http.Request) (interface{}, error) {
+	var req addHostsToTeamRequest
+
+	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
+		return nil, err
+	}
+	return req, nil
 }
